+++ frontend/src/components/(game)/game/index.tsx
@@ -8,6 +8,7 @@ import { SelectionProvider } from "@/contexts/(game)/selection";
 import { SearchContextProvider } from "@/contexts/(game)/search";
 import { AnimatePresence } from "framer-motion";
 import { Keybinds } from "../keybinds";
+import { ImportContextProvider } from "@/contexts/(game)/import";
 
 export type ShahrazadProps = {
     game: ShahrazadGame;
@@ -42,12 +43,14 @@ export default function Game(props: ShahrazadProps) {
             <SelectionProvider>
                 <ShahrazadDND>
                     <SearchContextProvider>
-                        <Keybinds />
-                        <AnimatePresence>
-                            <div className="mx-4 w-ful h-ful flex flex-col gap-4">
-                                {playmat_components}
-                            </div>
-                        </AnimatePresence>
+                        <ImportContextProvider>
+                            <Keybinds />
+                            <AnimatePresence>
+                                <div className="mx-4 w-ful h-ful flex flex-col gap-4">
+                                    {playmat_components}
+                                </div>
+                            </AnimatePresence>
+                        </ImportContextProvider>
                     </SearchContextProvider>
                 </ShahrazadDND>
             </SelectionProvider>
